+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Call skipWaiting() by default (can be disabled through flag).
 - The generated offline URLs and the SW registration is changed to relative URLS,
   we can put the application in any directory.
+- DynamicCache evicts older entries first (previously it was random).
 
 ## 0.0.1